refactor: made some name changes in all files
