--- src/ExpenseTracker/.editorconfig
@@ -1,151 +1,0 @@
-SA1629
-[*.{cs,vb}]
-#### Naming styles ####
-
-# Naming rules
-
-dotnet_naming_rule.interface_should_be_begins_with_i.severity = suggestion
-dotnet_naming_rule.interface_should_be_begins_with_i.symbols = interface
-dotnet_naming_rule.interface_should_be_begins_with_i.style = begins_with_i
-
-dotnet_naming_rule.types_should_be_pascal_case.severity = suggestion
-dotnet_naming_rule.types_should_be_pascal_case.symbols = types
-dotnet_naming_rule.types_should_be_pascal_case.style = pascal_case
-
-dotnet_naming_rule.non_field_members_should_be_pascal_case.severity = suggestion
-dotnet_naming_rule.non_field_members_should_be_pascal_case.symbols = non_field_members
-dotnet_naming_rule.non_field_members_should_be_pascal_case.style = pascal_case
-
-# Symbol specifications
-
-dotnet_naming_symbols.interface.applicable_kinds = interface
-dotnet_naming_symbols.interface.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.interface.required_modifiers = 
-
-dotnet_naming_symbols.types.applicable_kinds = class, struct, interface, enum
-dotnet_naming_symbols.types.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.types.required_modifiers = 
-
-dotnet_naming_symbols.non_field_members.applicable_kinds = property, event, method
-dotnet_naming_symbols.non_field_members.applicable_accessibilities = public, internal, private, protected, protected_internal, private_protected
-dotnet_naming_symbols.non_field_members.required_modifiers = 
-
-# Naming styles
-
-dotnet_naming_style.begins_with_i.required_prefix = I
-dotnet_naming_style.begins_with_i.required_suffix = 
-dotnet_naming_style.begins_with_i.word_separator = 
-dotnet_naming_style.begins_with_i.capitalization = pascal_case
-
-dotnet_naming_style.pascal_case.required_prefix = 
-dotnet_naming_style.pascal_case.required_suffix = 
-dotnet_naming_style.pascal_case.word_separator = 
-dotnet_naming_style.pascal_case.capitalization = pascal_case
-
-dotnet_naming_style.pascal_case.required_prefix = 
-dotnet_naming_style.pascal_case.required_suffix = 
-dotnet_naming_style.pascal_case.word_separator = 
-dotnet_naming_style.pascal_case.capitalization = pascal_case
-dotnet_style_operator_placement_when_wrapping = beginning_of_line
-tab_width = 4
-indent_size = 4
-end_of_line = crlf
-dotnet_style_coalesce_expression = true:suggestion
-dotnet_style_null_propagation = true:suggestion
-dotnet_style_prefer_is_null_check_over_reference_equality_method = true:suggestion
-dotnet_style_prefer_auto_properties = true:silent
-dotnet_style_object_initializer = true:suggestion
-dotnet_style_collection_initializer = true:suggestion
-
-[*.cs]
-csharp_indent_labels = one_less_than_current
-csharp_using_directive_placement = outside_namespace:silent
-csharp_prefer_simple_using_statement = true:suggestion
-csharp_prefer_braces = true:silent
-csharp_style_namespace_declarations = block_scoped:silent
-csharp_style_prefer_method_group_conversion = true:silent
-csharp_style_prefer_top_level_statements = true:silent
-csharp_style_expression_bodied_methods = false:silent
-csharp_style_expression_bodied_constructors = false:silent
-csharp_style_expression_bodied_operators = false:silent
-csharp_style_expression_bodied_properties = true:silent
-csharp_style_expression_bodied_indexers = true:silent
-csharp_style_expression_bodied_accessors = true:silent
-csharp_style_expression_bodied_lambdas = true:silent
-csharp_style_expression_bodied_local_functions = false:silent
-csharp_style_allow_embedded_statements_on_same_line_experimental = true:silent
-dotnet_diagnostic.SA1001.severity = error
-dotnet_diagnostic.SA1004.severity = error
-dotnet_diagnostic.SA1106.severity = error
-dotnet_diagnostic.SA1112.severity = error
-dotnet_diagnostic.SA1120.severity = error
-dotnet_diagnostic.SA1128.severity = error
-dotnet_diagnostic.SA1137.severity = error
-dotnet_diagnostic.SA1201.severity = error
-dotnet_diagnostic.SA1202.severity = error
-dotnet_diagnostic.SA1203.severity = error
-dotnet_diagnostic.SA1207.severity = error
-dotnet_diagnostic.SA1210.severity = error
-dotnet_diagnostic.SA1300.severity = error
-dotnet_diagnostic.SA1302.severity = error
-dotnet_diagnostic.SA1303.severity = error
-dotnet_diagnostic.SA1304.severity = error
-dotnet_diagnostic.SA1306.severity = error
-dotnet_diagnostic.SA1310.severity = error
-dotnet_diagnostic.SA1312.severity = error
-dotnet_diagnostic.SA1313.severity = error
-dotnet_diagnostic.SA1314.severity = error
-dotnet_diagnostic.SX1309.severity = error
-dotnet_diagnostic.SA1400.severity = error
-dotnet_diagnostic.SA1401.severity = error
-dotnet_diagnostic.SA1403.severity = error
-dotnet_diagnostic.SA1500.severity = error
-dotnet_diagnostic.SA1503.severity = error
-dotnet_diagnostic.SA1505.severity = error
-dotnet_diagnostic.SA1506.severity = error
-dotnet_diagnostic.SA1507.severity = error
-dotnet_diagnostic.SA1508.severity = error
-dotnet_diagnostic.SA1509.severity = error
-dotnet_diagnostic.SA1517.severity = error
-dotnet_diagnostic.SA1600.severity = error
-dotnet_diagnostic.SA1601.severity = error
-dotnet_diagnostic.SA1602.severity = error
-dotnet_diagnostic.SA1604.severity = error
-dotnet_diagnostic.SA1605.severity = error
-dotnet_diagnostic.SA1606.severity = error
-dotnet_diagnostic.SA1608.severity = error
-dotnet_diagnostic.SA1609.severity = error
-dotnet_diagnostic.SA1610.severity = error
-dotnet_diagnostic.SA1613.severity = error
-dotnet_diagnostic.SA1651.severity = error
-dotnet_diagnostic.SA1649.severity = error
-dotnet_diagnostic.SA1648.severity = error
-dotnet_diagnostic.SA1643.severity = error
-dotnet_diagnostic.SA1642.severity = error
-dotnet_diagnostic.SA1641.severity = error
-dotnet_diagnostic.SA1640.severity = error
-dotnet_diagnostic.SA1639.severity = error
-dotnet_diagnostic.SA1638.severity = error
-dotnet_diagnostic.SA1637.severity = error
-dotnet_diagnostic.SA1636.severity = silent
-dotnet_diagnostic.SA1635.severity = error
-dotnet_diagnostic.SA1634.severity = error
-dotnet_diagnostic.SA1627.severity = error
-dotnet_diagnostic.SA1626.severity = error
-dotnet_diagnostic.SA1625.severity = error
-dotnet_diagnostic.SA1624.severity = error
-dotnet_diagnostic.SA1611.severity = error
-dotnet_diagnostic.SA1612.severity = error
-dotnet_diagnostic.SA1614.severity = error
-dotnet_diagnostic.SA1615.severity = error
-dotnet_diagnostic.SA1616.severity = error
-dotnet_diagnostic.SA1617.severity = error
-dotnet_diagnostic.SA1618.severity = error
-dotnet_diagnostic.SA1619.severity = error
-dotnet_diagnostic.SA1620.severity = error
-dotnet_diagnostic.SA1621.severity = error
-dotnet_diagnostic.SA1622.severity = error
-dotnet_diagnostic.SA1629.severity = silent
-dotnet_diagnostic.SA1309.severity = silent
-dotnet_diagnostic.SA1633.severity = silent
-dotnet_diagnostic.SA1200.severity = silent
