+++ src/LINQ/Utilities/Utility.cs
@@ -1,11 +1,12 @@
-﻿using LINQ.Models;
+﻿using ConsoleTables;
+using LINQ.Models;
 
-namespace LINQ.Helpers;
+namespace LINQ.Utilities;
 
 /// <summary>
-/// Have helper function related to assignment.
+/// Have utilities function related to assignment.
 /// </summary>
-public class Helper
+public static class Utility
 {
     /// <summary>
     /// Add dummy products to the list.
@@ -81,5 +82,25 @@ public class Helper
         suppliers.Add(new Supplier { SupplierId = "SUP028", SupplierName = "PouchPro", ProductId = "028" });
         suppliers.Add(new Supplier { SupplierId = "SUP029", SupplierName = "ClipTech", ProductId = "029" });
         suppliers.Add(new Supplier { SupplierId = "SUP030", SupplierName = "HolderHub", ProductId = "030" });
+    }
+
+    /// <summary>
+    /// Print the products as table.
+    /// </summary>
+    /// <param name="products">Products list to print.</param>
+    public static void PrintProducts(List<Product> products)
+    {
+        ConsoleTable productTable = new ("Id", "Product", "Price", "Category");
+        foreach (Product product in products)
+        {
+            string?[] fields = new string[4];
+            fields[0] = product?.ID;
+            fields[1] = product?.Name;
+            fields[2] = product?.Price.ToString();
+            fields[3] = product?.Category;
+            productTable.AddRow(fields);
+        }
+
+        productTable.Write();
     }
 }
